Added training saving and dataset splitting
The dataset now gets split into training and validation sets, the ratio can be adjusted at DATA_SPLIT_RATIO. Added model saving, since before every time you ran training it was running an instance of the model. It created a catdog.nn file after training and looks for it when running, will train  a new instance if the file is not present. Removed the Vector struct since it was unused.